Localize use of dummy credentials
This is an attempt to fix https://github.com/localstack/serverless-localstack/issues/19

serverless-localstack should not change the AWS credentials environment variables,
as they will override the credentials configured by serverless.yml.

A better approach is to set the credentials directly when updating the AWS config.

I have verified that this fixes #19, but I cannot verify that it did not
break the feature added in https://github.com/localstack/serverless-localstack/pull/13
since that pull request did not include any tests.